Migrate Query.fieldListStats
